README: Recommend --debug in addition to --single
By default, unit-threaded buffers and hides stdout/stderr output. This
becomes a problem when an error is thrown in a unit test in a
Vibe.d task. Normally, uncaught throwables are printed to stderr by a
catch clause in vibe.core.task.TaskFiber.run, however, that output
would be hidden by unit-threaded by default. Because vibe-core calls
abort right after attempting to print the exception, we thus never get
to see it.

The solution is to use --debug in addition to --single, which disables
output buffering in unit-threaded.