Don't even query Bugzilla for non-Bugzilla projects
Follow-up to #272.

Apply the is-a-Bugzilla-using-project check to the code which parses
Bugzilla references and queries details for each from Bugzilla, too.

The test suite fallout is larger this time because we need to patch
the repo slug in the other-org tests, which means we can't reuse the
existing payloads (at least not without introducing a "read this other
payload, do a search&replace, and serve it instead" mechanism).